my contribution to BugDay... the 'integer' type as specified by W3C should be at least 64-bits wide. From now on MonetDB/XQuery conforms to the specs (regarding integers). (it used to only support integers in 32-bits).
My testweb is fairly green (FC4, 64bit), (greener than the testweb),
but i'm worried some new bugs will pop up due to this fix...

+++ pathfinder/compiler/core/core.c
@@ -215,7 +215,7 @@ PFcore_var (PFvar_t *v)
  * @return a new core tree node
  */
 PFcnode_t *
-PFcore_num (int i)
+PFcore_num (long long int i)
 {
     PFcnode_t *core;
